(zsh) Rename leading dots from config; reconfigure p10k
Since leading dots are supposed to represent hidden files, they should
not be present in the dotfiles repo as they aren't meant to be hidden in
the repo.

Use vi keybindings for zsh.

Reconfigure p10k to use two-line prompt; change prompt character in
insert mode to large right chevron.